New attempt at a beam-element container
This is supposed to replace the NS(Blocks) structure once it's clear that it
works for both OpenCL and CUDA and that the necessary API is in place.

Currently, it is very rudamentarily implemented, i.e. the API is not finished yet.

TODO: Flesh out the API
TODO: Write and apply unit tests!!!!