sixtracklib/common: add new version of the unified block-descriptor
This block descriptor is used to describe the memory layout and offsets
for all types of elements, including beam elements and particles. It also
contains a store_info field which either gives the id of the associated
mapping header or an per-block determined alignment

All offsets are in absolute offsts relative to the begin of a flat memory block.
This is a change compared to the previous implementation where particles and
beam elements had been serialized differently (because Particles where themselves
a container whereas beam elements where members of a container, i.e. a block)

Now both particles and beam elements can be serialized into their respective
containers