sixtracklib/common: Expanded NS(Blocks) API with a specialized unserialize function
The NS(Blocks_unserialize_without_remapping) function works like the already existing
NS(Blocks_unserialize) functions but only returns if no memory remapping has to be
performed. If the need for a memory remapping is detected, no action is performed and
an error is returned.

This is intended to be used in parallel where the remapping is not performed by all
processes/threads and the final unserialization should not alter the mapping. Until
now, we silently assumed that this would be the behaviour but with using this function
failures and problems can be detected.