common: works-around inconsistencies between the code-bases (temporary)
This keeps a minimal version of the context submodule around while the
parts of sixtracklib that depend on it (i.e. NS(ClContext), etc.) are
migrated to common/control and the hierachie therein

Note: so far, this change is only applied to the CMakeLists file. The
original common/context headers and sources are kept and will be removed in
a subsequent step