cuda: attempts to fix issue with uint64_t atomicOr
This attempts to addess comment #125 which fails on cc60
There seems to be an issue of typedef conflicts between host code and device
code on CUDA (most likely due to the use of uint64_t in common/definitions.h)

This is a work-around which should work. The better fix is to transit
the SIXTRL_(U)INT(xy) macros to use the proper native types and not to rely
on uint64_t etc. being consistent on the device side