tests/opencl: temporarily relax tolerance for track job unit test
- Intel OpenCL implementations based on the OneAPI seem to perform some
additional numerical optimization even with fastmath disabled. This results
in differences in the magnitude of 1e-12 in this test, potentially due to
fused muliply-add or similar differences in the calculations

This changes relax the comparison tolerances to allow the tests to pass, which
is not ideal.

TODO: create an issue to address this problem