sixtracklib/common: updates behaviour of interpolated space-charge map
By request of users, the map now checks whether the abscissa value for
which an interpolation should be calculated is within bounds (as expected)
or outside of the range sampled in the abscissa values

If the range is exceeded, 0 is returned rather than an assertion or error
being propagated. This is a deviation from the previous assumption, that
the values *always* have to be within bounds.

The performance impact for conforming values should be minimal (
one floating point multiplication, one addition, two comparisons,
one bitwise AND) and branch divergence should only occur in very exceptionally
few cases. Given the asymptotic behaviour of real-world profiles, this
should be a pretty sensible fall-back and avoids hard to debug / trace
errors if evaluation very close to the boundaries is required and
particles could stray outside of the bounds for example by rounding errors.

NOTE: we reserve the right to revert this change in case performance
is, unexpectedly, serverly affected. Please report if any unexpected
side-effects are associated with this change